Limit the results of the carbon query to only the current range
The carbon query will return all cached datapoints. This includes values
that are outside of your current query. When the carbonlink attempts to
put these values in the returned array, an index error gets thrown. This
causes a lot of errors that get thrown when datapoints are cached and
the requested range does not match the timestamps of the cached
datapoints.

Teach the carbon query how to limit which metrics get inserted into the
cache based on their timestamps.